Make iOS Drop-in platform API easier to navigate
